# openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/schedule/reservations": {
      "get": {
        "operationId": "ScheduleController_getMentorReservationList",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "페이지 번호",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "페이지당 항목 수",
            "schema": {
              "default": 10,
              "example": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "등록된 세션 목록",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MentorReservationListResponseDto"
                  }
                }
              }
            }
          },
          "500": {
            "description": "등록된 세션 목록을 찾을 수 없습니다."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "멘토 예약 목록 조회",
        "tags": [
          "Schedule"
        ]
      }
    },
    "/schedule/reservations/{id}": {
      "get": {
        "operationId": "ScheduleController_getMentorReservationDetail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "예약 상세 정보 반환",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MentorReservationDetailResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "멘토 예약 상세 조회",
        "tags": [
          "Schedule"
        ]
      }
    },
...
    "/reservation/available-times/{id}": {
      "get": {
        "operationId": "ReservationController_getAvailableTimes",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "required": true,
            "in": "query",
            "schema": {
              "example": "2025-07-08",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "예약 가능한 시간 목록을 반환합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "availableTime": [
                      {
                        "startTime": "14:00",
                        "endTime": "15:00"
                      },
                      {
                        "startTime": "15:00",
                        "endTime": "16:00"
                      }
                    ]
                  }
                }
              }
            }
          },
          "404": {
            "description": "세션 또는 멘토를 찾을 수 없습니다."
          },
          "500": {
            "description": "멘토링 예약 가능 시간 조회 중 오류가 발생했습니다."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "멘토링 예약 가능 시간 조회",
        "tags": [
          "Reservation"
        ]
      }
    },
    "/reservation/available-days/{id}": {
      "get": {
        "operationId": "ReservationController_getAvailableDays",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "예약 가능한 요일 리스트",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "availableDays": [
                      "MONDAY",
                      "WEDNESDAY",
                      "FRIDAY"
                    ]
                  }
                }
              }
            }
          },
          "404": {
            "description": "해당 멘토의 스케줄이 존재하지 않습니다."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "멘토링 예약 가능 날짜 조회",
        "tags": [
          "Reservation"
        ]
      }
    },
    "/reservation": {
      "post": {
        "operationId": "ReservationController_createReservation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReservationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "예약이 등록되고 결제를 유도합니다."
          },
          "404": {
            "description": "세션 또는 유저 정보를 찾을 수 없습니다."
          },
          "409": {
            "description": "해당 시간은 이미 예약되었습니다."
          },
          "500": {
            "description": "멘토링 예약 중 오류가 발생했습니다."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "멘토링 예약 등록",
        "tags": [
          "Reservation"
        ]
      }
    },
    "/reservation/my": {
      "get": {
        "operationId": "ReservationController_getMyReservations",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "페이지 번호 (기본값: 1)",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "페이지당 개수 (기본값: 10)",
            "schema": {
              "default": 10,
              "example": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "멘티 예약 내역을 조회합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReservationItemDto"
                  }
                }
              }
            }
          },
          "500": {
            "description": "예약 내역 조회 중 오류가 발생했습니다."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "멘티 진행중인 예약 내역 조회",
        "tags": [
          "Reservation"
        ]
      }
    },
    "/reservation/past": {
      "get": {
        "operationId": "ReservationController_getMyClearReservations",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "페이지 번호 (기본값: 1)",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "페이지당 개수 (기본값: 10)",
            "schema": {
              "default": 10,
              "example": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "멘티 예약 내역을 조회합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReservationClearItemDto"
                  }
                }
              }
            }
          },
          "500": {
            "description": "예약 내역 조회 중 오류가 발생했습니다."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "멘티 완료된 예약 내역 조회",
        "tags": [
          "Reservation"
        ]
      }
    },
    "/reservation/done/{orderId}": {
      "get": {
        "operationId": "ReservationController_donePayment",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "결제 완료 후 예약 확정 정보",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DonePaymentResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "결제 완료 페이지 - 예약 확정",
        "tags": [
          "Reservation"
        ]
      }
    },
    "/reservation/joinRoom/{roomId}": {
      "get": {
        "operationId": "ReservationController_joinRoom",
        "parameters": [
          {
            "name": "roomId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "입장 가능한 경우 meetingUrl 반환",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JoinRoomResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "입장 권한이 없거나 아직 시간이 안 됨"
          },
          "404": {
            "description": "예약을 찾을 수 없음"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "멘토링 방 입장",
        "tags": [
          "Reservation"
        ]
      }
    },
...
  },
  "info": {
    "title": "커넥트 api문서",
    "description": "커넥트 개발을 위한 api문서 \n\n🔗 OpenAPI 명세 다운로드: /openapi-spec.json`,",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "JWT Bearer Token"
      }
    },
    "schemas": {
...
      "MentorReservationListResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "예약 ID",
            "example": "b2a9c1ef-27f4-4c02-ae34-0fcecc30c9f1"
          },
          "title": {
            "type": "string",
            "description": "세션 제목",
            "example": "React 입문 강의"
          },
          "date": {
            "type": "string",
            "description": "예약날짜",
            "example": "2025-07-27"
          },
          "startTime": {
            "type": "string",
            "description": "시작 시간",
            "example": "14:00:00"
          },
          "endTime": {
            "type": "string",
            "description": "종료 시간",
            "example": "15:00:00"
          },
          "status": {
            "type": "string",
            "description": "예약 상태",
            "example": "pending",
            "enum": [
              "pending",
              "confirmed",
              "rejected"
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "예약 생성 시간",
            "example": "2025-07-27T14:32:00.000Z"
          },
          "menteeName": {
            "type": "string",
            "description": "멘티 이름",
            "example": "김철수"
          },
          "menteeEmail": {
            "type": "string",
            "description": "멘티 이메일",
            "example": "mentee@example.com"
          },
          "menteePhone": {
            "type": "string",
            "description": "멘티 전화번호",
            "example": "010-1234-5678"
          }
        },
        "required": [
          "id",
          "title",
          "date",
          "startTime",
          "endTime",
          "status",
          "createdAt",
          "menteeName",
          "menteeEmail",
          "menteePhone"
        ]
      },
      "MentorReservationDetailResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "예약 ID",
            "example": "b2a9c1ef-27f4-4c02-ae34-0fcecc30c9f1"
          },
          "title": {
            "type": "string",
            "description": "세션 제목",
            "example": "React 입문 강의"
          },
          "date": {
            "type": "string",
            "description": "예약 날짜",
            "example": "2025-07-30"
          },
          "startTime": {
            "type": "string",
            "description": "시작 시간",
            "example": "14:00:00"
          },
          "endTime": {
            "type": "string",
            "description": "종료 시간",
            "example": "15:00:00"
          },
          "question": {
            "type": "string",
            "description": "멘티가 남긴 사전 질문",
            "example": "포트폴리오에 뭐 넣어야 할까요?"
          },
          "rejectReason": {
            "type": "string",
            "description": "거절 사유",
            "example": "요청 시간에 일정이 겹칩니다."
          },
          "status": {
            "type": "string",
            "description": "예약 상태",
            "example": "pending",
            "enum": [
              "pending",
              "confirmed",
              "rejected"
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "예약 생성 시간",
            "example": "2025-07-27T14:32:00.000Z"
          },
          "menteeName": {
            "type": "string",
            "description": "멘티 이름",
            "example": "김철수"
          },
          "menteeEmail": {
            "type": "string",
            "description": "멘티 이메일",
            "example": "mentee@example.com"
          },
          "menteePhone": {
            "type": "string",
            "description": "멘티 전화번호",
            "example": "010-1234-5678"
          }
        },
        "required": [
          "id",
          "title",
          "date",
          "startTime",
          "endTime",
          "question",
          "status",
          "createdAt",
          "menteeName",
          "menteeEmail",
          "menteePhone"
        ]
      },
      "CreateReservationDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "example": "2025-07-08",
            "description": "예약 날짜 (YYYY-MM-DD)"
          },
          "startTime": {
            "type": "string",
            "example": "10:00",
            "description": "시작 시간 (HH:mm)"
          },
          "endTime": {
            "type": "string",
            "example": "11:00",
            "description": "종료 시간 (HH:mm)"
          },
          "sessionId": {
            "type": "string",
            "example": "uuid-string",
            "description": "멘토링 세션 ID"
          },
          "question": {
            "type": "string",
            "example": "React를 학습할 때 어떤 순서로 접근해야 하나요?",
            "description": "사전 질문"
          }
        },
        "required": [
          "date",
          "startTime",
          "endTime",
          "sessionId",
          "question"
        ]
      },
      "ReservationItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "a1b2c3d4-5678-90ef-ghij-klmnopqrstuv"
          },
          "date": {
            "type": "string",
            "example": "2025-07-10"
          },
          "startTime": {
            "type": "string",
            "example": "14:00"
          },
          "endTime": {
            "type": "string",
            "example": "15:00"
          },
          "status": {
            "type": "string",
            "example": "confirmed",
            "description": "예약 상태"
          },
          "sessionTitle": {
            "type": "string",
            "example": "포트폴리오 피드백"
          },
          "mentorName": {
            "type": "string",
            "example": "김멘토"
          },
          "roomId": {
            "type": "string",
            "example": "uuid"
          },
          "canEnter": {
            "type": "string",
            "example": "waiting",
            "enum": [
              "waiting",
              "progress",
              "closed"
            ],
            "default": "waiting"
          }
        },
        "required": [
          "id",
          "date",
          "startTime",
          "endTime",
          "status",
          "sessionTitle",
          "mentorName",
          "roomId",
          "canEnter"
        ]
      },
      "ReservationClearItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "a1b2c3d4-5678-90ef-ghij-klmnopqrstuv"
          },
          "date": {
            "type": "string",
            "example": "2025-07-10"
          },
          "startTime": {
            "type": "string",
            "example": "14:00"
          },
          "endTime": {
            "type": "string",
            "example": "15:00"
          },
          "status": {
            "type": "string",
            "example": "completed",
            "description": "예약 상태"
          },
          "sessionTitle": {
            "type": "string",
            "example": "포트폴리오 피드백"
          },
          "mentorName": {
            "type": "string",
            "example": "김멘토"
          },
          "reviewWritten": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "id",
          "date",
          "startTime",
          "endTime",
          "status",
          "sessionTitle",
          "mentorName",
          "reviewWritten"
        ]
      },
      "DonePaymentResponseDto": {
        "type": "object",
        "properties": {
          "reservationId": {
            "type": "string",
            "description": "예약 ID",
            "example": "6f4c1c3d-89ab-4c3d-b3f2-1f9a3a2b7e6f"
          },
          "mentorName": {
            "type": "string",
            "description": "멘토 이름",
            "example": "홍길동"
          },
          "date": {
            "type": "string",
            "description": "예약 날짜",
            "example": "2025-08-21"
          },
          "startTime": {
            "type": "string",
            "description": "시작 시간",
            "example": "10:00"
          },
          "endTime": {
            "type": "string",
            "description": "종료 시간",
            "example": "11:00"
          },
          "meetingUrl": {
            "type": "string",
            "description": "화상 미팅 URL (아직 생성 안됐으면 null)",
            "example": "https://meet.konnect.com/abc123",
            "nullable": true
          }
        },
        "required": [
          "reservationId",
          "mentorName",
          "date",
          "startTime",
          "endTime"
        ]
      },
      "JoinRoomResponseDto": {
        "type": "object",
        "properties": {
          "reservationId": {
            "type": "string",
            "example": "a6f8e12c-1234-4567-8901-abcdef123456"
          },
          "meetingUrl": {
            "type": "string",
            "example": "https://konnect.com/room/550e8400-e29b-41d4-a716-446655440000"
          }
        },
        "required": [
          "reservationId",
          "meetingUrl"
        ]
      },
...
    }
  }
}